#26 Has been Added Validation for LeaveAllocation Features

+++ HR.LeaveManagement.Application/Features/LeaveAllocation/Commands/Delete/DeleteLeaveAllocationCommandHandler.cs
@@ -15,6 +15,12 @@ public class DeleteLeaveAllocationCommandHandler : IRequestHandler<DeleteLeaveAl
 
     public async Task<Unit> Handle(DeleteLeaveAllocationCommand request, CancellationToken cancellationToken)
     {
+        DeleteLeaveAllocationCommandValidation validation = new();
+        var validationResult = await validation.ValidateAsync(request);
+
+        if (!validationResult.IsValid)
+            throw new BadRequestException("Invalid Request", validationResult);
+
         var deleteLeaveAllocation = await _repository.GetByIdAsync(request.Id) ?? throw new NotFoundException(nameof(Domain.LeaveAllocation), request.Id);
 
         await _repository.DeleteAsync(deleteLeaveAllocation);
